Added list items and create item functions

+++ GreenFoxFinalHomework/Services/Interfaces/IListingService.cs
@@ -5,7 +5,7 @@ namespace GreenFoxFinalHomework.Services.Interfaces
 {
     public interface IListingService
     {
-        Item CreateItem(string name, string description, string photoUrl, double startingPrice, int userId);
+        Item CreateItem(string name, string description, string photoUrl, int startingPrice, int userId);
         Item ViewItem(int id);
         List<Item> ListItems();
         List<Item> ListItems(int howMany);
